Change type of pop variable from float to Py_ssize_t

+++ skimage/filter/rank/_core8.pxd
@@ -12,7 +12,7 @@ cdef np.uint8_t is_in_mask(Py_ssize_t rows, Py_ssize_t cols,
 
 # 8 bit core kernel receives extra information about data inferior and superior
 # percentiles
-cdef void _core8(np.uint8_t kernel(Py_ssize_t *, float, np.uint8_t, float,
+cdef void _core8(np.uint8_t kernel(Py_ssize_t *, Py_ssize_t, np.uint8_t, float,
                                    float, Py_ssize_t, Py_ssize_t),
                  np.ndarray[np.uint8_t, ndim=2] image,
                  np.ndarray[np.uint8_t, ndim=2] selem,
