+++ skimage/filter/rank/_crank16_bilateral.pyx
@@ -13,7 +13,7 @@ from skimage.filter.rank._core16 cimport _core16
 # -----------------------------------------------------------------
 
 
-cdef inline np.uint16_t kernel_mean(Py_ssize_t * histo, float pop,
+cdef inline np.uint16_t kernel_mean(Py_ssize_t * histo, Py_ssize_t pop,
                                     np.uint16_t g, Py_ssize_t bitdepth,
                                     Py_ssize_t maxbin, Py_ssize_t midbin,
                                     float p0, float p1,
@@ -35,7 +35,7 @@ cdef inline np.uint16_t kernel_mean(Py_ssize_t * histo, float pop,
         return <np.uint16_t>(0)
 
 
-cdef inline np.uint16_t kernel_pop(Py_ssize_t * histo, float pop,
+cdef inline np.uint16_t kernel_pop(Py_ssize_t * histo, Py_ssize_t pop,
                                    np.uint16_t g, Py_ssize_t bitdepth,
                                    Py_ssize_t maxbin, Py_ssize_t midbin,
                                    float p0, float p1,
